Did a lot of stuff read description
I set up the search bar to be able to change between teacher and
student, though I didn’t know what to do with the edit button so I
stuck it next to the search bar. I fixed the header, I think. I also
set up an idea for a simple algorithm that checks certain things to
determine whether it’s a room or id. I’m still trying to plan out how
it will check whether it’s a first or a last name, but we can figure
that out later. Solid design for the presentation.